chore(api-builder): linkDevGuide -> linkDocs;linkDocs can default the lang closes #1849
1) Rename linkDevGuide to linkDocs

Rename the specialized API docs `@linkDevGuide` tag to `@linkDocs`
since it is meant to be used to refer to all angular.io docs, not just
the developer guide.

2) let linkDocs links resolve to proper language

The old linkDevGuide required URIs to be given in the form
'ts/latest/guide/architecture' which forced all language versions of
the API pages to refer to the ts versions of the doc pages they linked
to.

This PR fixes that problem. URIs are now given in the form
'guide/architecture' and then rendered properly for each language. Of
course, when needed, it is still possible to give a language specific
URI.

chore(api-builder): minor code cleanup